requirements.txt
fastapi
python-decouple
requests
uvicorn